Adjusted welcomeMessage functions. Added new editEmbed functions (add-field and edit-image not available)
