+++ src/commands/embeds/remove-embed-image.js
@@ -5,7 +5,7 @@ const {
 } = require("discord.js");
 
 module.exports = {
-  name: "edit-embed-title",
+  name: "remove-embed-image",
   description: "Descripción",
   options: [
     {
@@ -17,12 +17,6 @@ module.exports = {
     {
       name: "mensaje",
       description: "Ingrese el ID del mensaje con el Embed a modificar.",
-      type: ApplicationCommandOptionType.String,
-      required: true,
-    },
-    {
-      name: "título",
-      description: "Ingrese un nuevo título para el Embed.",
       type: ApplicationCommandOptionType.String,
       required: true,
     },
@@ -44,21 +38,17 @@ module.exports = {
 
       const channelId = interaction.options.get("canal").value;
       const messageId = interaction.options.get("mensaje").value;
-      const newTitle = interaction.options.get("título").value;
 
       const channel = client.channels.cache.get(channelId);
       const message = await channel.messages.fetch(messageId);
       const originalEmbed = message.embeds[0];
 
       const newEmbed = new EmbedBuilder()
-        .setTitle(newTitle)
+        .setTitle(originalEmbed.tlitle)
         .setDescription(originalEmbed.description)
         .setColor(originalEmbed.color)
         .setTimestamp();
 
-      if (originalEmbed.image) {
-        newEmbed.setImage(originalEmbed.image);
-      }
       if (originalEmbed.thumbnail) {
         newEmbed.setThumbnail(`${originalEmbed.thumbnail}`);
       }
